requirements.txt
cohere
cohere-haystack
haystack-ai
haystack-experimental
chroma-haystack
python-dotenv
faiss-cpu
tika
qdrant-haystack
pypdf
-e .
